services/enemy-marking/requirements.txt
# services/enemy-marking/requirements.txt
fastapi
uvicorn[standard]
pydantic
kafka-python